move npu_vm_runtime.hpp to utils/vm
