+++ src/plugins/intel_npu/src/CMakeLists.txt
@@ -13,7 +13,6 @@ add_subdirectory(ops)
 
 if (ENABLE_NPU_PLUGIN_ENGINE)
     add_subdirectory(common)
-    add_subdirectory(vm)
     add_subdirectory(compiler_adapter)
     add_subdirectory(backend)
     add_subdirectory(plugin)
