+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/vm/dynamic_arguments.hpp
@@ -10,7 +10,7 @@
 #include <string>
 #include <vector>
 
-#include "intel_npu/runtime/npu_vm_runtime.hpp"
+#include "intel_npu/utils/vm/npu_vm_runtime.hpp"
 #include "openvino/core/except.hpp"
 #include "openvino/core/shape.hpp"
 #include "openvino/runtime/itensor.hpp"
