+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/vm/npu_vm_runtime.hpp
