+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/vm/npu_vm_runtime_api.hpp
@@ -7,7 +7,7 @@
 #include <memory>
 #include <string_view>
 
-#include "intel_npu/runtime/npu_vm_runtime.hpp"
+#include "npu_vm_runtime.hpp"
 #include "openvino/core/except.hpp"
 
 namespace intel_npu {
