--- src/plugins/intel_npu/src/vm/CMakeLists.txt
@@ -1,6 +1,0 @@
-#
-# Copyright (C) 2026 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_subdirectory(runtime)
