--- src/plugins/intel_npu/src/vm/runtime/CMakeLists.txt
@@ -1,16 +1,0 @@
-#
-# Copyright (C) 2026 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-set(TARGET_NAME npu_vm_runtime_api)
-
-add_library(${TARGET_NAME} INTERFACE)
-add_library(openvino::npu_vm_runtime_api ALIAS ${TARGET_NAME})
-
-target_include_directories(${TARGET_NAME} INTERFACE
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-)
-
-ov_add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
